requirements.txt
aiofiles==25.1.0
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.6
distro==1.9.0
fastapi==0.135.3
filetype==1.2.0
google-auth==2.49.1
google-genai==1.70.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
kombu==5.6.2
langchain==1.2.15
langchain-core==1.2.26
langchain-google-genai==4.2.1
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.25
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.4
openai==2.30.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pgvector==0.4.2
pillow==12.2.0
prompt_toolkit==3.0.52
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
PyYAML==6.0.3
redis==7.4.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.43.0
vine==5.1.0
wcwidth==0.6.0
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
